NSUInteger for cache

+++ pantomime-lib/Framework/Pantomime/Extensions/CacheRecord.h
@@ -10,12 +10,12 @@
 
 @interface CacheRecord : NSObject
 
-@property (nonatomic) unsigned int date;
-@property (nonatomic) unsigned int flags;
-@property (nonatomic) unsigned int position;   // For mbox based cache files
-@property (nonatomic) unsigned int size;
-@property (nonatomic) unsigned int imap_uid;   // For IMAP
-@property (nonatomic) char *filename;          // For maildir base cache files
+@property (nonatomic) NSUInteger date;
+@property (nonatomic) NSUInteger flags;
+@property (nonatomic) NSUInteger position;   // For mbox based cache files
+@property (nonatomic) NSUInteger size;
+@property (nonatomic) NSUInteger imap_uid;   // For IMAP
+@property (nonatomic) char *filename;        // For maildir base cache files
 @property (nonatomic, strong) NSString *pop3_uid;      // For POP3
 @property (nonatomic, strong) NSData *from;
 @property (nonatomic, strong) NSData *in_reply_to;
